chore: 调整 parseUrl 方法的暴露路径
