+++ packages/taro-runtime/src/index.ts
@@ -10,7 +10,7 @@ export { History } from './bom/history'
 export { Location } from './bom/location'
 export { nav as navigator } from './bom/navigator'
 export { caf as cancelAnimationFrame, now, raf as requestAnimationFrame } from './bom/raf'
-export { URL } from './bom/URL'
+export { parseUrl,URL } from './bom/URL'
 export { URLSearchParams } from './bom/URLSearchParams'
 export { history, location, window } from './bom/window'
 // dom
@@ -24,7 +24,6 @@ export { SVGElement } from './dom/svg'
 export { TaroText } from './dom/text'
 export { MutationObserver } from './dom-external/mutation-observer'
 // others
-export { parseUrl } from './bom/location'
 export { Current, getCurrentInstance } from './current'
 export { eventSource } from './dom/event-source'
 export {
